Makes version required for bundle download task
With the recent changes to the new demo server there is no longer a
"bundle-latest.zip" file. Instead we are using the year of the bundle,
i.e. "bundle-2016.zip".

This change enforces the need to supply the version via `ENV.fetch`
which will now throw an exception if not supplied.

Fixes #426